+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Important note: All settings have been reset to default values.**
 
+- Added: Option to hide overview entries without balance changes
 - Added: Profile settings with account profile and character profile selection
 - Updated: deDE, enUS, ruRU localization
 - Minor code adjustments